Current Version of Robot.cpp
I'm currently writing up the install instructions and figured it would be easiest to create zip files for the user to download and compile and run. These don't have the .o and .exe versions of the server or the robot, but I will include build instructions in the file. I'll upload the install instructions later when I'm more or less done, I may have a couple of things to discuss with you all though. I can add in that they should copy in whichever core/completion/challenge robot.cpp and robot.hpp files instead of the one in the zip file if that would work.

If you have a different idea as to how to have the user download everything, let me know in the comments on this commit/pull request and we can discuss it.